requirements.txt
# A股金融年报ChatBI系统依赖包
# ===============================

# 核心框架
qwen-agent>=0.0.5
dashscope>=1.17.0

# 数据处理
pandas>=2.0.0
numpy>=1.24.0
sqlalchemy>=2.0.0
mysql-connector-python>=8.0.0

# 数据可视化
matplotlib>=3.7.0
seaborn>=0.12.0

# 机器学习和时间序列分析
scikit-learn>=1.3.0
statsmodels>=0.14.0
prophet>=1.1.0

# Web界面
gradio>=4.0.0

# API和网络请求
requests>=2.30.0
tavily-python>=0.3.0

# 数据采集
baostock>=0.8.8

# 工具库
python-dateutil>=2.8.0
pytz>=2023.3
